+++ packages/ai-constructs/src/conversation/runtime/types.ts
@@ -26,6 +26,14 @@ export type ConversationMessageContentBlock =
         // Upstream (Appsync) may send images in a form of Base64 encoded strings
         source: { bytes: string };
       };
+      // These are needed so that union with other content block types works.
+      // See https://docs.aws.amazon.com/AWSJavaScriptSDK/v3/latest/Package/-aws-sdk-client-bedrock-runtime/TypeAlias/ContentBlock/.
+      text?: never;
+      document?: never;
+      toolUse?: never;
+      toolResult?: never;
+      guardContent?: never;
+      $unknown?: never;
     };
 
 export type ToolDefinition<TJSONSchema extends JSONSchema = JSONSchema> = {
@@ -39,6 +47,7 @@ export type ToolDefinition<TJSONSchema extends JSONSchema = JSONSchema> = {
 export type ConversationTurnEvent = {
   conversationId: string;
   currentMessageId: string;
+  streamResponse?: boolean;
   responseMutation: {
     name: string;
     inputTypeName: string;
@@ -56,14 +65,8 @@ export type ConversationTurnEvent = {
     };
   };
   request: {
-    headers: {
-      authorization: string;
-    };
+    headers: Record<string, string>;
   };
-  /**
-   * @deprecated This field is going to be removed in upcoming releases.
-   */
-  messages?: Array<ConversationMessage>;
   messageHistoryQuery: {
     getQueryName: string;
     getQueryInputTypeName: string;
@@ -91,3 +94,43 @@ export type ExecutableTool<
 > = ToolDefinition<TJSONSchema> & {
   execute: (input: TToolInput) => Promise<ToolResultContentBlock>;
 };
+
+export type StreamingResponseChunk = {
+  // always required
+  conversationId: string;
+  associatedUserMessageId: string;
+  contentBlockIndex: number;
+} & (
+  | {
+      // text chunk
+      contentBlockText: string;
+      contentBlockDeltaIndex: number;
+      contentBlockDoneAtIndex?: never;
+      contentBlockToolUse?: never;
+      stopReason?: never;
+    }
+  | {
+      // end of block. applicable to text blocks
+      contentBlockDoneAtIndex: number;
+      contentBlockText?: never;
+      contentBlockDeltaIndex?: never;
+      contentBlockToolUse?: never;
+      stopReason?: never;
+    }
+  | {
+      // tool use
+      contentBlockToolUse: string; // serialized json with full tool use block
+      contentBlockDoneAtIndex?: never;
+      contentBlockText?: never;
+      contentBlockDeltaIndex?: never;
+      stopReason?: never;
+    }
+  | {
+      // turn complete
+      stopReason: string;
+      contentBlockDoneAtIndex?: never;
+      contentBlockText?: never;
+      contentBlockDeltaIndex?: never;
+      contentBlockToolUse?: never;
+    }
+);