+++ packages/integration-tests/src/test-project-setup/conversation_handler_project.ts
@@ -7,10 +7,7 @@ import { AmplifyClient } from '@aws-sdk/client-amplify';
 import { BackendIdentifier } from '@aws-amplify/plugin-types';
 import { InvokeCommand, LambdaClient } from '@aws-sdk/client-lambda';
 import { DeployedResourcesFinder } from '../find_deployed_resource.js';
-import {
-  ConversationMessage,
-  ConversationTurnEvent,
-} from '@aws-amplify/ai-constructs/conversation/runtime';
+import { ConversationTurnEvent } from '@aws-amplify/ai-constructs/conversation/runtime';
 import { randomUUID } from 'crypto';
 import { generateClientConfig } from '@aws-amplify/client-config';
 import { AmplifyAuthCredentialsFactory } from '../amplify_auth_credentials_factory.js';
@@ -30,10 +27,11 @@ import { NormalizedCacheObject } from '@apollo/client';
 import {
   bedrockModelId,
   expectedTemperatureInDataToolScenario,
-  expectedTemperatureInProgrammaticToolScenario,
+  expectedTemperaturesInProgrammaticToolScenario,
 } from '../test-projects/conversation-handler/amplify/constants.js';
 import { resolve } from 'path';
 import { fileURLToPath } from 'url';
+import * as bedrock from '@aws-sdk/client-bedrock-runtime';
 
 // TODO: this is a work around
 // it seems like as of amplify v6 , some of the code only runs in the browser ...
@@ -51,35 +49,43 @@ type ConversationTurnAppSyncResponse = {
   content: string;
 };
 
+type ConversationMessage = {
+  role: 'user' | 'assistant';
+  content: Array<ConversationMessageContentBlock>;
+};
+
+type ConversationMessageContentBlock =
+  | bedrock.ContentBlock
+  | {
+      image: Omit<bedrock.ImageBlock, 'source'> & {
+        // Upstream (Appsync) may send images in a form of Base64 encoded strings
+        source: { bytes: string };
+      };
+      // These are needed so that union with other content block types works.
+      // See https://docs.aws.amazon.com/AWSJavaScriptSDK/v3/latest/Package/-aws-sdk-client-bedrock-runtime/TypeAlias/ContentBlock/.
+      text?: never;
+      document?: never;
+      toolUse?: never;
+      toolResult?: never;
+      guardContent?: never;
+      $unknown?: never;
+    };
+
 type CreateConversationMessageChatInput = ConversationMessage & {
   conversationId: string;
   id: string;
   associatedUserMessageId?: string;
 };
 
-const commonEventProperties = {
-  responseMutation: {
-    name: 'createConversationMessageAssistantResponse',
-    inputTypeName: 'CreateConversationMessageAssistantResponseInput',
-    selectionSet: [
-      'id',
-      'conversationId',
-      'content',
-      'owner',
-      'createdAt',
-      'updatedAt',
-    ].join('\n'),
-  },
-  messageHistoryQuery: {
-    getQueryName: 'getConversationMessageChat',
-    getQueryInputTypeName: 'ID',
-    listQueryName: 'listConversationMessageChats',
-    listQueryInputTypeName: 'ModelConversationMessageChatFilterInput',
-  },
-  modelConfiguration: {
-    modelId: bedrockModelId,
-    systemPrompt: 'You are helpful bot.',
-  },
+type ConversationTurnAppSyncResponseChunk = {
+  conversationId: string;
+  associatedUserMessageId: string;
+  contentBlockIndex: number;
+  contentBlockText?: string;
+  contentBlockDeltaIndex?: number;
+  contentBlockDoneAtIndex?: number;
+  contentBlockToolUse?: string;
+  stopReason?: string;
 };
 
 /**
@@ -204,8 +210,6 @@ class ConversationHandlerTestProject extends TestProjectBase {
       authenticatedUserCredentials.accessToken,
       clientConfig.data.url,
       apolloClient,
-      // Does not use message history lookup.
-      // This case should be removed when event.messages field is removed.
       false
     );
 
@@ -222,7 +226,7 @@ class ConversationHandlerTestProject extends TestProjectBase {
       authenticatedUserCredentials.accessToken,
       clientConfig.data.url,
       apolloClient,
-      true,
+      false,
       // Simulate eventual consistency
       true
     );
@@ -231,21 +235,48 @@ class ConversationHandlerTestProject extends TestProjectBase {
       backendId,
       authenticatedUserCredentials.accessToken,
       clientConfig.data.url,
-      apolloClient
+      apolloClient,
+      false
+    );
+
+    await this.assertCustomConversationHandlerCanExecuteTurn(
+      backendId,
+      authenticatedUserCredentials.accessToken,
+      clientConfig.data.url,
+      apolloClient,
+      true
     );
 
     await this.assertDefaultConversationHandlerCanExecuteTurnWithDataTool(
       backendId,
       authenticatedUserCredentials.accessToken,
       clientConfig.data.url,
-      apolloClient
+      apolloClient,
+      false
+    );
+
+    await this.assertDefaultConversationHandlerCanExecuteTurnWithDataTool(
+      backendId,
+      authenticatedUserCredentials.accessToken,
+      clientConfig.data.url,
+      apolloClient,
+      true
     );
 
     await this.assertDefaultConversationHandlerCanExecuteTurnWithClientTool(
       backendId,
       authenticatedUserCredentials.accessToken,
       clientConfig.data.url,
-      apolloClient
+      apolloClient,
+      false
+    );
+
+    await this.assertDefaultConversationHandlerCanExecuteTurnWithClientTool(
+      backendId,
+      authenticatedUserCredentials.accessToken,
+      clientConfig.data.url,
+      apolloClient,
+      true
     );
 
     await this.assertDefaultConversationHandlerCanExecuteTurnWithImage(
@@ -270,7 +301,7 @@ class ConversationHandlerTestProject extends TestProjectBase {
     accessToken: string,
     graphqlApiEndpoint: string,
     apolloClient: ApolloClient<NormalizedCacheObject>,
-    useMessageHistory: boolean,
+    streamResponse: boolean,
     withoutMessageAvailableInTheMessageList = false
   ): Promise<void> => {
     const defaultConversationHandlerFunction = (
@@ -300,38 +331,23 @@ class ConversationHandlerTestProject extends TestProjectBase {
       request: {
         headers: { authorization: accessToken },
       },
-      ...commonEventProperties,
+      ...this.getCommonEventProperties(streamResponse),
     };
 
-    if (useMessageHistory) {
-      if (withoutMessageAvailableInTheMessageList) {
-        // This tricks conversation handler to think that message is not available in the list.
-        // I.e. it simulates eventually consistency read at list operation where item is not yet visible.
-        // In this case handler should fall back to lookup by current message id.
-        message.conversationId = randomUUID().toString();
-      }
-      await this.insertMessage(apolloClient, message);
-    } else {
-      event.messageHistoryQuery = {
-        getQueryName: '',
-        getQueryInputTypeName: '',
-        listQueryName: '',
-        listQueryInputTypeName: '',
-      };
-      event.messages = [
-        {
-          role: message.role,
-          content: message.content,
-        },
-      ];
+    if (withoutMessageAvailableInTheMessageList) {
+      // This tricks conversation handler to think that message is not available in the list.
+      // I.e. it simulates eventually consistency read at list operation where item is not yet visible.
+      // In this case handler should fall back to lookup by current message id.
+      message.conversationId = randomUUID().toString();
     }
+    await this.insertMessage(apolloClient, message);
 
-    const response = await this.executeConversationTurn(
+    const responseContent = await this.executeConversationTurn(
       event,
       defaultConversationHandlerFunction,
       apolloClient
     );
-    assert.match(response.content, /3\.14/);
+    assert.match(responseContent, /3\.14/);
   };
 
   private assertDefaultConversationHandlerCanExecuteTurnWithImage = async (
@@ -339,7 +355,7 @@ class ConversationHandlerTestProject extends TestProjectBase {
     accessToken: string,
     graphqlApiEndpoint: string,
     apolloClient: ApolloClient<NormalizedCacheObject>,
-    useMessageHistory: boolean
+    streamResponse: boolean
   ): Promise<void> => {
     const defaultConversationHandlerFunction = (
       await this.resourceFinder.findByBackendIdentifier(
@@ -388,39 +404,25 @@ class ConversationHandlerTestProject extends TestProjectBase {
       request: {
         headers: { authorization: accessToken },
       },
-      ...commonEventProperties,
+      ...this.getCommonEventProperties(streamResponse),
     };
-    if (useMessageHistory) {
-      await this.insertMessage(apolloClient, message);
-    } else {
-      event.messageHistoryQuery = {
-        getQueryName: '',
-        getQueryInputTypeName: '',
-        listQueryName: '',
-        listQueryInputTypeName: '',
-      };
-      event.messages = [
-        {
-          role: message.role,
-          content: message.content,
-        },
-      ];
-    }
-    const response = await this.executeConversationTurn(
+    await this.insertMessage(apolloClient, message);
+    const responseContent = await this.executeConversationTurn(
       event,
       defaultConversationHandlerFunction,
       apolloClient
     );
     // The image contains a logo of AWS. Responses may vary, but they should always contain statements below.
-    assert.match(response.content, /logo/);
-    assert.match(response.content, /(aws)|(AWS)|(Amazon Web Services)/);
+    assert.match(responseContent, /logo/);
+    assert.match(responseContent, /(aws)|(AWS)|(Amazon Web Services)/);
   };
 
   private assertDefaultConversationHandlerCanExecuteTurnWithDataTool = async (
     backendId: BackendIdentifier,
     accessToken: string,
     graphqlApiEndpoint: string,
-    apolloClient: ApolloClient<NormalizedCacheObject>
+    apolloClient: ApolloClient<NormalizedCacheObject>,
+    streamResponse: boolean
   ): Promise<void> => {
     const defaultConversationHandlerFunction = (
       await this.resourceFinder.findByBackendIdentifier(
@@ -477,16 +479,16 @@ class ConversationHandlerTestProject extends TestProjectBase {
           },
         ],
       },
-      ...commonEventProperties,
+      ...this.getCommonEventProperties(streamResponse),
     };
-    const response = await this.executeConversationTurn(
+    const responseContent = await this.executeConversationTurn(
       event,
       defaultConversationHandlerFunction,
       apolloClient
     );
     // Assert that tool was used. I.e. that LLM used value returned by the tool.
     assert.match(
-      response.content,
+      responseContent,
       new RegExp(expectedTemperatureInDataToolScenario.toString())
     );
   };
@@ -495,7 +497,8 @@ class ConversationHandlerTestProject extends TestProjectBase {
     backendId: BackendIdentifier,
     accessToken: string,
     graphqlApiEndpoint: string,
-    apolloClient: ApolloClient<NormalizedCacheObject>
+    apolloClient: ApolloClient<NormalizedCacheObject>,
+    streamResponse: boolean
   ): Promise<void> => {
     const defaultConversationHandlerFunction = (
       await this.resourceFinder.findByBackendIdentifier(
@@ -545,9 +548,9 @@ class ConversationHandlerTestProject extends TestProjectBase {
           },
         ],
       },
-      ...commonEventProperties,
+      ...this.getCommonEventProperties(streamResponse),
     };
-    const response = await this.executeConversationTurn(
+    const responseContent = await this.executeConversationTurn(
       event,
       defaultConversationHandlerFunction,
       apolloClient
@@ -555,17 +558,157 @@ class ConversationHandlerTestProject extends TestProjectBase {
     // Assert that tool use content blocks are emitted in case LLM selects client tool.
     // The content blocks are string serialized, but not as a proper JSON,
     // hence string matching is employed below to detect some signals that tool use blocks kinds were emitted.
-    assert.match(response.content, /toolUse/);
-    assert.match(response.content, /toolUseId/);
+    assert.match(responseContent, /toolUse/);
+    assert.match(responseContent, /toolUseId/);
     // Assert that LLM attempts to pass parameter when asking for tool use.
-    assert.match(response.content, /"city":"Seattle"/);
+    assert.match(responseContent, /"city":"Seattle"/);
+  };
+
+  private executeConversationTurn = async (
+    event: ConversationTurnEvent,
+    functionName: string,
+    apolloClient: ApolloClient<NormalizedCacheObject>
+  ): Promise<string> => {
+    await this.lambdaClient.send(
+      new InvokeCommand({
+        FunctionName: functionName,
+        Payload: Buffer.from(JSON.stringify(event)),
+      })
+    );
+
+    // assert that response came back
+    if (event.streamResponse) {
+      let nextToken: string | undefined;
+      const chunks: Array<ConversationTurnAppSyncResponseChunk> = [];
+      do {
+        const queryResult = await apolloClient.query<{
+          listConversationMessageAssistantStreamingResponses: {
+            items: Array<ConversationTurnAppSyncResponseChunk>;
+            nextToken: string | undefined;
+          };
+        }>({
+          query: gql`
+            query ListMessageChunks(
+              $conversationId: ID
+              $associatedUserMessageId: ID
+              $nextToken: String
+            ) {
+              listConversationMessageAssistantStreamingResponses(
+                limit: 1000
+                nextToken: $nextToken
+                filter: {
+                  conversationId: { eq: $conversationId }
+                  associatedUserMessageId: { eq: $associatedUserMessageId }
+                }
+              ) {
+                items {
+                  associatedUserMessageId
+                  contentBlockDeltaIndex
+                  contentBlockDoneAtIndex
+                  contentBlockIndex
+                  contentBlockText
+                  contentBlockToolUse
+                  conversationId
+                  createdAt
+                  id
+                  owner
+                  stopReason
+                  updatedAt
+                }
+                nextToken
+              }
+            }
+          `,
+          variables: {
+            conversationId: event.conversationId,
+            associatedUserMessageId: event.currentMessageId,
+            nextToken,
+          },
+          fetchPolicy: 'no-cache',
+        });
+        nextToken =
+          queryResult.data.listConversationMessageAssistantStreamingResponses
+            .nextToken;
+        chunks.push(
+          ...queryResult.data.listConversationMessageAssistantStreamingResponses
+            .items
+        );
+      } while (nextToken);
+
+      assert.ok(chunks);
+
+      chunks.sort((a, b) => {
+        // This is very simplified sort by message,block and delta indexes;
+        let aValue = 1000 * 1000 * a.contentBlockIndex;
+        if (a.contentBlockDeltaIndex) {
+          aValue += a.contentBlockDeltaIndex;
+        }
+        let bValue = 1000 * 1000 * b.contentBlockIndex;
+        if (b.contentBlockDeltaIndex) {
+          bValue += b.contentBlockDeltaIndex;
+        }
+        return aValue - bValue;
+      });
+
+      const content = chunks.reduce((accumulated, current) => {
+        if (current.contentBlockText) {
+          accumulated += current.contentBlockText;
+        }
+        if (current.contentBlockToolUse) {
+          accumulated += current.contentBlockToolUse;
+        }
+        return accumulated;
+      }, '');
+
+      return content;
+    }
+    const queryResult = await apolloClient.query<{
+      listConversationMessageAssistantResponses: {
+        items: Array<ConversationTurnAppSyncResponse>;
+      };
+    }>({
+      query: gql`
+        query ListMessage($conversationId: ID, $associatedUserMessageId: ID) {
+          listConversationMessageAssistantResponses(
+            filter: {
+              conversationId: { eq: $conversationId }
+              associatedUserMessageId: { eq: $associatedUserMessageId }
+            }
+          ) {
+            items {
+              conversationId
+              id
+              updatedAt
+              createdAt
+              content
+              associatedUserMessageId
+            }
+            nextToken
+          }
+        }
+      `,
+      variables: {
+        conversationId: event.conversationId,
+        associatedUserMessageId: event.currentMessageId,
+      },
+      fetchPolicy: 'no-cache',
+    });
+    assert.strictEqual(
+      1,
+      queryResult.data.listConversationMessageAssistantResponses.items.length
+    );
+    const response =
+      queryResult.data.listConversationMessageAssistantResponses.items[0];
+    assert.ok(response.content);
+    return response.content;
   };
 
   private assertCustomConversationHandlerCanExecuteTurn = async (
     backendId: BackendIdentifier,
     accessToken: string,
     graphqlApiEndpoint: string,
-    apolloClient: ApolloClient<NormalizedCacheObject>
+    apolloClient: ApolloClient<NormalizedCacheObject>,
+    streamResponse: boolean
   ): Promise<void> => {
     const customConversationHandlerFunction = (
       await this.resourceFinder.findByBackendIdentifier(
@@ -581,7 +724,7 @@ class ConversationHandlerTestProject extends TestProjectBase {
       role: 'user',
       content: [
         {
-          text: 'What is the temperature in Seattle?',
+          text: 'What is the temperature in Seattle and Boston?',
         },
       ],
     };
@@ -595,61 +738,26 @@ class ConversationHandlerTestProject extends TestProjectBase {
       request: {
         headers: { authorization: accessToken },
       },
-      ...commonEventProperties,
+      ...this.getCommonEventProperties(streamResponse),
     };
-    const response = await this.executeConversationTurn(
+    const responseContent = await this.executeConversationTurn(
       event,
       customConversationHandlerFunction,
       apolloClient
     );
     // Assert that tool was used. I.e. LLM used value provided by the tool.
     assert.match(
-      response.content,
-      new RegExp(expectedTemperatureInProgrammaticToolScenario.toString())
+      responseContent,
+      new RegExp(
+        expectedTemperaturesInProgrammaticToolScenario.Seattle.toString()
+      )
     );
-  };
-
-  private executeConversationTurn = async (
-    event: ConversationTurnEvent,
-    functionName: string,
-    apolloClient: ApolloClient<NormalizedCacheObject>
-  ): Promise<ConversationTurnAppSyncResponse> => {
-    await this.lambdaClient.send(
-      new InvokeCommand({
-        FunctionName: functionName,
-        Payload: Buffer.from(JSON.stringify(event)),
-      })
+    assert.match(
+      responseContent,
+      new RegExp(
+        expectedTemperaturesInProgrammaticToolScenario.Boston.toString()
+      )
     );
-
-    // assert that response came back
-
-    const queryResult = await apolloClient.query<{
-      listConversationMessageAssistantResponses: {
-        items: Array<ConversationTurnAppSyncResponse>;
-      };
-    }>({
-      query: gql`
-        query ListMessages {
-          listConversationMessageAssistantResponses(limit: 1000) {
-            items {
-              conversationId
-              id
-              updatedAt
-              createdAt
-              content
-              associatedUserMessageId
-            }
-          }
-        }
-      `,
-      fetchPolicy: 'no-cache',
-    });
-    const response =
-      queryResult.data.listConversationMessageAssistantResponses.items.find(
-        (item) => item.associatedUserMessageId === event.currentMessageId
-      );
-    assert.ok(response);
-    return response;
   };
 
   private insertMessage = async (
@@ -668,5 +776,43 @@ class ConversationHandlerTestProject extends TestProjectBase {
         input: message,
       },
     });
+  };
+
+  private getCommonEventProperties = (streamResponse: boolean) => {
+    const responseMutation = streamResponse
+      ? {
+          name: 'createConversationMessageAssistantStreamingResponse',
+          inputTypeName:
+            'CreateConversationMessageAssistantStreamingResponseInput',
+          selectionSet: ['id', 'conversationId', 'createdAt', 'updatedAt'].join(
+            '\n'
+          ),
+        }
+      : {
+          name: 'createConversationMessageAssistantResponse',
+          inputTypeName: 'CreateConversationMessageAssistantResponseInput',
+          selectionSet: [
+            'id',
+            'conversationId',
+            'content',
+            'owner',
+            'createdAt',
+            'updatedAt',
+          ].join('\n'),
+        };
+    return {
+      streamResponse,
+      responseMutation,
+      messageHistoryQuery: {
+        getQueryName: 'getConversationMessageChat',
+        getQueryInputTypeName: 'ID',
+        listQueryName: 'listConversationMessageChats',
+        listQueryInputTypeName: 'ModelConversationMessageChatFilterInput',
+      },
+      modelConfiguration: {
+        modelId: bedrockModelId,
+        systemPrompt: 'You are helpful bot.',
+      },
+    };
   };
 }