+++ packages/backend-function/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aws-amplify/backend-function
 
+## 1.7.0
+
+### Minor Changes
+
+- 4720412: Add minify option to defineFunction
+
+## 1.6.0
+
+### Minor Changes
+
+- f5d0ab4: adds support to reference existing layers in defineFunction
+
 ## 1.5.0
 
 ### Minor Changes