+++ packages/backend/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @aws-amplify/backend
 
+## 1.5.0
+
+### Minor Changes
+
+- 4720412: Add minify option to defineFunction
+
+### Patch Changes
+
+- Updated dependencies [f87cc87]
+- Updated dependencies [4720412]
+  - @aws-amplify/backend-secret@1.1.4
+  - @aws-amplify/backend-function@1.7.0
+
+## 1.4.0
+
+### Minor Changes
+
+- f5d0ab4: adds support to reference existing layers in defineFunction
+
+### Patch Changes
+
+- Updated dependencies [f5d0ab4]
+  - @aws-amplify/backend-function@1.6.0
+
 ## 1.3.2
 
 ### Patch Changes